Many of the drawfuncitons are fucked by the Renderer changes. Because i named the renderers screen width to width. Needs fixing

+++ GUITesting/Point2D.h
@@ -2,11 +2,12 @@
 class Point2D {
 public:
 	Point2D(double xC = 0, double yC = 0) : x(xC), y(yC) {}
+	//Point2D(Point3D a) : x(a.X()), y(a.Y()){}
 private:
 	double x, y;
 public:
 	double X() { return x; }
 	double Y() { return y; }
-	void X(int a) { x = a; }
-	void Y(int a) { y = a; }
+	void X(double a) { x = a; }
+	void Y(double a) { y = a; }
 };
